Moves LCW code to common.
