+++ common/lcw.cpp
