+++ common/lcw.h
