+++ common/lcwcompa.asm
