+++ common/lcwuncmp.cpp
